backend/requirements.txt
flask
flask-cors
ffmpeg-python
openai-whisper
